backend/requirements.txt
langchain
langchain-community
langchain-openai
langchain-pinecone
pinecone-client
python-dotenv